requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
pillow
opencv-python
numpy
pytesseract
ultralytics
easyocr
